Move TextEditingArea and its *Actions interfaces into view package
