+++ richtextfx/src/integrationTest/java/org/fxmisc/richtext/api/HitTests.java
@@ -7,7 +7,7 @@ import javafx.geometry.Point2D;
 import javafx.geometry.Pos;
 import javafx.stage.Stage;
 import org.fxmisc.richtext.InlineCssTextAreaAppTest;
-import org.fxmisc.richtext.model.NavigationActions;
+import org.fxmisc.richtext.NavigationActions;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
