+++ richtextfx/src/main/java/org/fxmisc/richtext/Selection.java
@@ -133,7 +133,7 @@ public interface Selection<PS, SEG, S> {
     /**
      * Selects the given range.
      *
-     * <p><b>Caution:</b> see {@link org.fxmisc.richtext.model.TextEditingArea#getAbsolutePosition(int, int)} to
+     * <p><b>Caution:</b> see {@link TextEditingArea#getAbsolutePosition(int, int)} to
      * know how the column index argument can affect the returned position.</p>
      */
     void selectRange(int startParagraphIndex, int startColPosition, int endParagraphIndex, int endColPosition);
